Solana settlement/settlement encoding (#4779)
# Description
<!--- Describe your changes to provide context for reviewers, including
why it is needed -->

# Changes
- Adds a basic `Solana` adapter struct (wrapping the previous
`SolanaRpc` value and the `program_id` value)
- Adds the `Settlement` domain type, which can create a transaction to
the Settlement program.
- it uses the `cow_settlement_client` whenever possible to not drift
from the program's cannon logic.
  
The two are not connected yet, but this will be exercised in the follow
up PRs, as well as wiring them w/ the `/settle` endpoint.

+++ crates/solana-driver/src/domain/solution.rs
@@ -24,7 +24,7 @@ pub struct Solution {
     /// Address lookup tables the interactions assume.
     pub address_lookup_tables: Vec<Pubkey>,
     /// Optional solver estimate of total settlement compute units.
-    pub cu_estimate: Option<u64>,
+    pub cu_estimate: Option<u32>,
 }
 
 /// A fulfillment of one auction order.
